app/requirements.txt
anyio==3.4.0
asgiref==3.4.1
azure-core==1.21.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.1
fonttools==4.28.5
h11==0.12.0
# httptools==0.2.0
idna==3.3
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
msrest==0.6.21
numpy==1.21.5
oauthlib==3.1.1
orjson==3.6.5
packaging==21.3
Pillow==8.4.0
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sniffio==1.2.0
starlette==0.16.0
threadpoolctl==3.0.0
typing-extensions==4.0.1
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
